Added functions as seperate classes
